chore(`errors`): split into `blockchain_info`, `info` & `balance`
